WeatherFlow station parser (#4)
* Add parser for local (UDP) WeatherFlow data.

Initial debugging/prototype parser for WeatherFlow UDP data. The
WeatherFlow Smart Weather Stations broadcast local weather data to a
UDP port with a well defined format.  The parser will listen for that
data and translate it into a format that the RainMachine can use.

* Report min/max temperature and min/max humidity.

Add report of daily temperature and humidity min/max values.

* Clean up comments.

Remove a bunch of the commented code and comments from the example.
Add some clarifying comments.

* Change parameter names.

It seems like parameters won't save if the names have spaces or
maybe '/' characters in them.

* Clean up unused variables.

* log message improvements.

also some minor changes in the thread loop condition.

* Handle the thread loop exiting gracefully.

Allow the thread to restart if it exits.

* Remove unused methods for int/float conversion.

* Add comment block with some explanation of how the parser works.

* Fix and enable code to run under pycharm.

This code block is strickly for testing under pycharm.

* Make debugging log messages use debug log.

Only makes sense to use the debug log method for debug log messages.