Updates to the WeatherFlow parser (#13)
* WF: change rain total from hour accumlation to day accumlation

* WF: Rewrite how data is collected and sent.

Based on feedback/discussions with Nicu Pavel the parser should be
sending accumulated average or accumulated total values, not hourly.
Only the last reported value for the day is saved.

Providing data this way actually simplifies the program. We simply
keep a running average/total in the UDP listener and report out the
current value at each parseInterval.

Note that I'm not convinced that a running average makes sense for
solar radiation (or wind speed).

* WF: Fix a couple of errors

 - parserData needed to be initialized as a two element array of dictds
 - datetime.datetime.now() should be only datetime.now
 - typeo in microsecond on make timestamp
 - missing 'self' on parserData assignment

* WF: More fixes

A couple more typos were corrected:
   RMParser should be RMParsers
   humidity was spelled wrong in one place

Add a version number/data string to the debug output to help with
identifying the version running.

Added comments around parserInternval

* WF: Fix temperature min/max and add min/max to debug log output.

* WF: Reset daily totals for rain and solar radiation.

* WF: Update to version 1.1.2

* WF: Send yesterday's value with timestamp set to start of day.

When sending the data for the previous day, set the time stamp to
the start of the day, 00:00:00, instead of using the last timestamp
from the day. Matches better with what the RainMachine is expecting.

* WF: Update to version 1.1.3